Move hooks files into the areas where they are used.
