+++ src/workspaces/container/WorkspaceContainer.test.ts
@@ -7,8 +7,8 @@ import { workspacesStore, workspaceStore } from 'src/libs/state';
 import { WorkspaceWrapper } from 'src/libs/workspace-utils';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultAzureWorkspace, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
+import { InitializedWorkspaceWrapper } from 'src/workspaces/container/state/useWorkspace';
 import { WorkspaceContainer } from 'src/workspaces/container/WorkspaceContainer';
-import { InitializedWorkspaceWrapper } from 'src/workspaces/hooks/useWorkspace';
 
 type NavExports = typeof import('src/libs/nav');
 jest.mock(
